Create README.md
CHADA TECH
Module Eight Journal Summarize the Project and what problem it was solving? 
In project one I worked the Chada Tech Clock. The client wanted to be able to view a 12- and 24-hour clock on their website in both formats simultaneously. The 12-hour clock should be in 12-hour format and should not exceed 12:59:59. The 24-hour clock should be displayed in 24-hour formation and does not exceed 23:59:59. Both clocks should display on the screen next to each other and have a menu to be able to add one hour, and one minute, add one second or exit the program.
What did you do particularly well? 
The was the first project in the course, I was not sure where to start. I had to use resources like google. YouTube and zybook for review. The flow chart was already given as part of the project. It made it easier to know the program before starting. See flow chart on next page.

 
This program could be enhanced by using the computer’s real time. This is something I will try after the course. I think the most challenging part of this project was to have both times to change simultaneously while keeping the desired format. In this case, I did a lot of trial and error. I didn’t have a specific method to use to solve the issue. I just kept trying until I finally got it. I also went back to zybook to practice functions but that’s really all I did to get it to work. I also find YouTube very helpful for this class. My biggest struggle in this class was pointer and linked list. I plan on practicing those while I’m taking next term off along with enhancing all my project.
I tried to make the program neat and maintainable by using functions and commenting each line. That way if I go back to it in the future, I wouldn’t have to spend so much time trying to figure out what I did.  I also used whitespace to keep my code readable.  I used variables that make sense. If someone were to read my code, they wouldn’t have to spend time trying to figure out what the variables in my code are. I think that really helped with making my code readable and maintainable.